+++ rust/operator-binary/src/main.rs
@@ -1,4 +1,4 @@
-use std::{os::unix::prelude::FileTypeExt, path::PathBuf};
+use std::{os::unix::prelude::FileTypeExt, path::PathBuf, pin::pin};
 
 use anyhow::Context;
 use clap::{crate_description, crate_version, Parser};
@@ -10,9 +10,7 @@ use futures::{FutureExt, TryStreamExt};
 use grpc::csi::v1::{
     controller_server::ControllerServer, identity_server::IdentityServer, node_server::NodeServer,
 };
-use stackable_operator::{
-    logging::TracingTarget, utils::cluster_info::KubernetesClusterInfoOpts, CustomResourceExt,
-};
+use stackable_operator::{cli::ProductOperatorRun, CustomResourceExt};
 use tokio::signal::unix::{signal, SignalKind};
 use tokio_stream::wrappers::UnixListenerStream;
 use tonic::transport::Server;
@@ -24,6 +22,7 @@ mod csi_server;
 mod external_crd;
 mod format;
 mod grpc;
+mod truststore_controller;
 mod utils;
 
 pub const APP_NAME: &str = "secret";
@@ -53,12 +52,8 @@ struct SecretOperatorRun {
     #[clap(long, env)]
     privileged: bool,
 
-    /// Tracing log collector system
-    #[arg(long, env, default_value_t, value_enum)]
-    pub tracing_target: TracingTarget,
-
-    #[command(flatten)]
-    pub cluster_info_opts: KubernetesClusterInfoOpts,
+    #[clap(flatten)]
+    common: ProductOperatorRun,
 }
 
 mod built_info {
@@ -71,13 +66,19 @@ async fn main() -> anyhow::Result<()> {
     match opts.cmd {
         stackable_operator::cli::Command::Crd => {
             crd::SecretClass::print_yaml_schema(built_info::PKG_VERSION)?;
+            crd::TrustStore::print_yaml_schema(built_info::PKG_VERSION)?;
         }
         stackable_operator::cli::Command::Run(SecretOperatorRun {
             csi_endpoint,
             node_name,
-            tracing_target,
             privileged,
-            cluster_info_opts,
+            common:
+                ProductOperatorRun {
+                    product_config: _,
+                    watch_namespace,
+                    tracing_target,
+                    cluster_info_opts,
+                },
         }) => {
             stackable_operator::logging::initialize_logging(
                 "SECRET_PROVISIONER_LOG",
@@ -105,7 +106,7 @@ async fn main() -> anyhow::Result<()> {
                 let _ = std::fs::remove_file(&csi_endpoint);
             }
             let mut sigterm = signal(SignalKind::terminate())?;
-            Server::builder()
+            let csi_server = pin!(Server::builder()
                 .add_service(
                     tonic_reflection::server::Builder::configure()
                         .include_reflection_service(true)
@@ -117,7 +118,7 @@ async fn main() -> anyhow::Result<()> {
                     client: client.clone(),
                 }))
                 .add_service(NodeServer::new(SecretProvisionerNode {
-                    client,
+                    client: client.clone(),
                     node_name,
                     privileged,
                 }))
@@ -127,8 +128,13 @@ async fn main() -> anyhow::Result<()> {
                     )
                     .map_ok(TonicUnixStream),
                     sigterm.recv().map(|_| ()),
-                )
-                .await?;
+                ));
+            let truststore_controller =
+                pin!(truststore_controller::start(&client, &watch_namespace).map(Ok));
+            futures::future::select(csi_server, truststore_controller)
+                .await
+                .factor_first()
+                .0?;
         }
     }
     Ok(())